refactor(voice): split voice.rs into voice/{settings,types,registry,providers,inference,mel,download}.rs
Behaviour-neutral move of the 3013-line voice.rs into a directory module.
Code was relocated verbatim by exact line ranges (zero non-blank lines
dropped); no logic, signatures, field names, or cfg attributes changed.
voice/audio.rs is untouched (byte-identical to HEAD).

Layout:
- mod.rs   - facade: shared imports (pub(super) use), all consts
             (pub(super)), mod decls, pub(crate) use re-exports preserving
             every crate::voice::* path, and the inline #[cfg(test)] mod tests.
- settings.rs - VoiceSettingsData + VoiceSettings + persistence.
- types.rs    - provider enums, payload structs, VoiceProvider/VoiceTranscriber traits.
- registry.rs - VoiceProviderRegistry struct + impl (recording orchestration).
- providers.rs- Platform/DistilWhisper providers, platform_* cfg helpers,
                ActiveDownloadGuard, enabled_key/model_status_key/distil_model_ready.
- inference.rs- candle/Whisper core (named inference to avoid colliding with the
                candle_transformers whisper alias): CandleBackend, transcriber,
                WhisperModel/Decoder, device/backend select+verify+probe, transcribe.
- mel.rs      - build_mel_filters, hz_to_mel, mel_to_hz.
- download.rs - download_distil_model.

Visibility widened to cross the new module boundaries (all minimal,
pub(super) only; nothing made bare pub that wasn't already):
- mod.rs: all shared `use` imports -> pub(super) use; all 11 consts -> pub(super);
  audio re-export block -> pub(super) use.
- types.rs: VoiceTranscriber trait -> pub(super).
- inference.rs: CandleBackend, CandleWhisperTranscriber, CandleBackendChecker,
  DefaultCandleBackendChecker, select_candle_device, ensure_candle_backend_ready,
  check_candle_backend_ready, transcribe_distil_whisper, timeout_error_message -> pub(super).
- providers.rs: PlatformVoiceProvider, DistilWhisperCandleProvider, ActiveDownloadGuard,
  enabled_key, model_status_key, distil_model_ready -> pub(super).
- download.rs: download_distil_model -> pub(super).
- mel.rs: build_mel_filters, hz_to_mel, mel_to_hz -> pub(super).

No boundary collapsed; coupling crossed cleanly via pub(super) re-exports.

check green: Tests 1480 passed (1480); Rust test result: ok. 318 passed; 0 failed.
